+++ NEWS.md
@@ -1,3 +1,7 @@
+# africamonitor 0.2.1
+- `am_pivot_wider()` supports factor variables as labels and does not check for uniqueness, making it a bit faster. 
+- `am_pivot_longer()` now by default codes series and labels columns as factors, which also yields speed improvements. 
+
 # africamonitor 0.2.0
 - Significant expansion of the database to >700 series with high Africa coverage, and computation of coverage statistics in the `am_series()` table, July 2022